% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateAMDominancy.R
\name{calculateAMDominancy}
\alias{calculateAMDominancy}
\title{Calculate A-M-Dominancy}
\usage{
calculateAMDominancy(ps, action, p.measures, show.info = TRUE)
}
\arguments{
\item{ps}{[\code{PreferenceSystem}]\cr
Preference System. Possibly calculated with
\code{\link{makeDecisionProblem}}, or initialized manually}

\item{action}{[\code{character}]\cr
The act for which it should be checked if it is
A-M-dominant.\cr
Must be one of the levels of the \code{action} variable
in the \code{data.frame} of the object \code{ps$df}.}

\item{p.measures}{[\code{matrix}]\cr
Matrix of probability vectors. Each row corresponds
to one probability vector.}

\item{show.info}{[\code{logical(1L)}]
Print info output to console? Defaults to \code{TRUE}.}
}
\value{
[\code{numeric()}]
  Vector of minimal optimal value of the objective functions as returned by
  \code{\link{calculateLocalAdmissibility}}.
}
\description{
Checks if an action is A-M-dominant for a given set of
  probability measures.
}
\references{
\insertRef{mainref}{dips}
}
