% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateGIE.R
\name{calculateGIE}
\alias{calculateGIE}
\title{Calculate Generalized Interval Expectation}
\usage{
calculateGIE(ps, delta, p.measures, action)
}
\arguments{
\item{ps}{[\code{dp}]\cr
Preference System calculated with\code{\link{makeDecisionProblem}}.}

\item{delta}{[\code{numeric(1L)}]\cr
granularity parameter. Must be between 0 and 1}

\item{p.measures}{[\code{list}]\cr
List of probability measures. Each entry must have exactly
\code{n.state}.\cr
Where \code{n.state} corresponds to the number
of levels the variable \code{state} has in the \code{data.frame}
of the object \code{ps$df}.}

\item{action}{[\code{character}]\cr
The act that the interval is calculated for.\cr
Must be one of the levels of the \code{action} variable
in the \code{data.frame} of the object \code{ps$df}.}
}
\value{
[\code{numeric(2L)}]\cr
  Lower and upper bounds of the interval.
}
\description{
Calculates the Generalized Interval Expectation for an act,
  on the basis of a consistent Preference System.
}
\references{
Christoph Jansen & Georg Schollmeyer & Thomas Augustin 2018\cr
Concepts for Decision Making under Severe
Uncertainty with Partial Ordinal and Partial
Cardinal Preferences.\cr
International Journal of Approximate Reasoning (V. 98)
}
