% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateLocalAdmissibility.R
\name{calculateLocalAdmissibility}
\alias{calculateLocalAdmissibility}
\title{Calculate Local-Admissibility}
\usage{
calculateLocalAdmissibility(ps, action1, action2, p.measures,
  keep.models = FALSE)
}
\arguments{
\item{ps}{[\code{dp}]\cr
Preference System calculated with \code{\link{makeDecisionProblem}}.}

\item{action1}{[\code{character}]\cr
The act that is preferred over action2.}

\item{action2}{[\code{character}]\cr
The second element of the tupel, where action2 is not as desireable
as action1.}

\item{p.measures}{[\code{matrix}]\cr
Matrix of probability vectors. Each row corresponds
to one probability vector.}

\item{keep.models}{[\code{logical(1L)}]\cr
Should all linear programs (e.g.: \code{\link{lp.object}})
be stored in the return object.
Can be memory consuming for bigger problems,
if set to \code{TRUE}. Defaults to \code{FALSE}.}
}
\value{
[\code{numeric(2L)}]\cr
  Optimal values of the set of linear programs in minimum and maximum form.
}
\description{
Checks if a tupel of actions, where action1 is preferred over action2,
  is an element of Preference Relations, as defined in Jansen et. al 2018
  (see references).
}
\references{
\insertRef{mainref}{dips}
}
