% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateProbabilityExtremePoints.R
\name{calculateProbabilityExtremePoints}
\alias{calculateProbabilityExtremePoints}
\title{Calculates Extreme Points for a Probability Model}
\usage{
calculateProbabilityExtremePoints(a1, b1, a2, b2)
}
\arguments{
\item{a1}{Numerical or character matrix for
inequality constraints.
If vector, treated as matrix with one row.}

\item{b1}{Numerical or character right
hand side vector for inequality constraints}

\item{a2}{Numerical or character matrix for equality constraints.
If vector, treated as matrix with one row.}

\item{b2}{Numerical or character right hand side vector
for equality constraints.}
}
\value{
[\code{matrix}]\cr
  Where each row corresponds to one probability vector.
}
\description{
Calculates Extreme points for a given Probability
  Model. Hereby, the model is given in form of
  (in-)equality constraints. See 'details'.
}
\details{
\code{calculateProbabilityExtremePoints} uses the \code{rcdd}
  package to calculate the extreme points. The unit simplex is already
  setup internally, so only the the probabilistic constraints
  need to be passed to \code{calculateProbabilityExtremePoints}.
  See \code{\link[rcdd]{makeH}} and the
  \href{https://cran.r-project.org/web/packages/rcdd/vignettes/vinny.pdf}{vignette}
  of the \code{rcdd} package for further details.
}
\references{
Christoph Jansen & Georg Schollmeyer & Thomas Augustin 2018\cr
Concepts for Decision Making under Severe
Uncertainty with Partial Ordinal and Partial
Cardinal Preferences.\cr
International Journal of Approximate Reasoning (V. 98)
}
