% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkPreferenceSystem.R
\name{checkPreferenceSystem}
\alias{checkPreferenceSystem}
\title{Check if an object is a Prefrence System}
\usage{
checkPreferenceSystem(ps)
}
\arguments{
\item{ps}{[\code{dp}]\cr
Preference System calculated with\code{\link{makeDecisionProblem}}.}
}
\value{
[\code{PreferenceSystem}]\cr
  If checks were succesful, the input is returned.
}
\description{
This function is more used internally to check
  if an object has the right structure to qualify
  as a Prefrence System. However, for users
  that want to set up their own relations R1 and R2
  this function can be useful to check if the
  object was constructed the right way.
}
