% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{costumer.ratings}
\alias{costumer.ratings}
\title{Costumer Ratings.}
\format{A data frame with 4 rows and 5 variables:
\describe{
  \item{earnings}{Expected gains}
  \item{s1}{Logical. Star 1}
  \item{s2}{Logical. Star 2}
  \item{s3}{Logical. Star 3}
  \item{product}{Factor with 2 levels, describing
  the different states of nature.}
  \item{packaging}{Factor with 2 levels, one for each type of packaging.}
}}
\usage{
costumer.ratings
}
\description{
Contains the data for a decision problem that was
used as an example in the Master Thesis that
was written about this package.\cr
Each row corresponds to one of the 4 alternatives.
}
\keyword{data}
