% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatePreferenceSystemConsistency.R
\name{getI}
\alias{getI}
\title{Get Indifference of Set}
\usage{
getI(x)
}
\arguments{
\item{x}{[\code{matrix}|\code{data.frame}]\cr
Matrix or data.frame with 2 or 4 columns to work on R1, R2 respectively.}
}
\value{
[\code{data.frame}]\cr
}
\description{
Calculates indifference part from a \code{matrix} or a \code{data.frame.}
}
