% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatePreferenceSystemConsistency.R
\name{getP}
\alias{getP}
\title{Get Strict Part Of Set}
\usage{
getP(x)
}
\arguments{
\item{x}{[\code{matrix}|\code{data.frame}]\cr
Matrix or data.frame with 2 or 4 columns to work on R1, R2 respectively.}
}
\value{
[\code{data.frame}]
}
\description{
Calculates strict part from a \code{matrix} or a \code{data.frame.}
}
