% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{job.offers}
\alias{job.offers}
\title{Job Offers.}
\format{A data frame with 12 rows and 8 variables:
\describe{
  \item{salary}{Offered wages per month}
  \item{b1}{Logical. Additional benefit: Overtime}
  \item{b2}{Logical. Additional benefit: Child care}
  \item{b3}{Logical. Additional benefit: Training}
  \item{b4}{Logical. Additional benefit: Prospects}
  \item{b5}{Logical. Additional benefit: Flexibility}
  \item{nature}{Factor with 4 levels, describing
  the different states of nature.}
  \item{job}{Factor with 4 levels, one for each job.}
}}
\usage{
job.offers
}
\description{
Contains the data for a decision problem that was
used as an example in the paper:\cr
Concepts for Decision Making under Severe
Uncertainty with Partial Ordinal and Partial
Cardinal Preferences.\cr
Each row corresponds to one of the 12 alternatives.
}
\references{
Christoph Jansen & Georg Schollmeyer & Thomas Augustin 2018\cr
Concepts for Decision Making under Severe
Uncertainty with Partial Ordinal and Partial
Cardinal Preferences.\cr
International Journal of Approximate Reasoning (V. 98)
}
\keyword{data}
