% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDecisionProblem.R
\name{makeDecisionProblem}
\alias{makeDecisionProblem}
\title{Construct Decision Problem}
\usage{
makeDecisionProblem(df, state, action, exclude = NULL)
}
\arguments{
\item{df}{[\code{data.frame}]\cr
Data from which the Decision Problem shall be derived.}

\item{state}{[\code{character(1)}]\cr
Name of the variable assigning the states of nature.}

\item{action}{[\code{character(1)}]\cr
Name of the variable assigning the acts to the data frame rows.}

\item{exclude}{[\code{character}]\cr
Variable names that should be excluded from the decision problem.
Default is \code{NULL}, meaning no variable is excluded.}
}
\value{
[\code{DecisionProblem}] With entries:\cr
  df: The original data frame\cr
  cardinal.information: List of cardinal utility for each
  combination of state and action.\cr
  ordinal.information: List of ordinal utility for each
  combination of state and action.
}
\description{
Constructs a Decision Problem from a data frame.
}
