% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePreferenceSystem.R
\name{makePreferenceSystem}
\alias{makePreferenceSystem}
\title{Construct A Preference System}
\usage{
makePreferenceSystem(dp)
}
\arguments{
\item{dp}{[\code{dp}]\cr
Decison Problem calculated with \code{\link{makeDecisionProblem}}.}
}
\value{
[\code{list}] With entries:\cr
  R1: Preorder on the acts.\cr
  R2: Preorder on R1.
}
\description{
Constructs a Preference System from a Decision Problem.
}
