% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGIE.R
\name{plotGIE}
\alias{plotGIE}
\title{Plot Generalized Interval Expectation}
\usage{
plotGIE(ps, deltas, p.measures, actions)
}
\arguments{
\item{ps}{[\code{dp}]\cr
Decison Problem calculated with \code{\link{makeDecisionProblem}}.}

\item{deltas}{[\code{numeric()}]\cr
Vector of granularity parameters. Each entry must be between 0 and 1.}

\item{p.measures}{[\code{list}]\cr
List of probability measures. Each entry must have exactly
\code{n.state}.\cr
Where \code{n.state} corresponds to the number
of levels the variable \code{state} has in the \code{data.frame}
of the object \code{ps$df}.}

\item{actions}{[\code{character}]\cr
The acts that the interval is calculated for.\cr
Must be one of the levels of the \code{action} variable
in the \code{data.frame} of the object \code{ps$df}.}
}
\value{
ggplot2 object.
}
\description{
Generates a plot visualizing Generalized Interval Expectation
  for a set of actions from a Preference System
  and a vector of granularities.
}
