% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateAAdmissibility.R
\name{calculateAAdmissibility}
\alias{calculateAAdmissibility}
\title{Construct A Preference System}
\usage{
calculateAAdmissibility(ps, action, p.measures)
}
\arguments{
\item{ps}{[\code{dp}]\cr
Decison Problem calculated with \code{\link{makeDecisionProblem}}.}

\item{action}{[\code{character}]\cr
The act that the interval is calculated for.\cr
Must be one of the levels of the \code{action} variable
in the \code{data.frame} of the object \code{ps$df}.}

\item{p.measures}{[\code{list}]\cr
List of probability measures. Each entry must have exactly
\code{n.state}.\cr
Where \code{n.state} corresponds to the number
of levels the variable \code{state} has in the \code{data.frame}
of the object \code{ps$df}.}
}
\value{
[\code{list}] With entries:\cr
  R1: Pre order on the acts.\cr
  R2: Preorder on R1.
}
\description{
Constructs a Preference System from a Decision Problem.
}
