% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateAAdmissibility.R
\name{calculateAAdmissibility}
\alias{calculateAAdmissibility}
\title{Calculate A-Admissibility}
\usage{
calculateAAdmissibility(ps, action, p.measures)
}
\arguments{
\item{ps}{[\code{dp}]\cr
Preference System calculated with \code{\link{makeDecisionProblem}}.}

\item{action}{[\code{character}]\cr
The act for which it should be checked if it is
A-Admissible.\cr
Must be one of the levels of the \code{action} variable
in the \code{data.frame} of the object \code{ps$df}.}

\item{p.measures}{[\code{matrix}]\cr
Matrix of probability vectors. Each row corresponds
to one probability vector.}
}
\value{
[\code{numeric}]\cr
  Optimal value of the objective function.
}
\description{
Checks if an action is A-Admissible for a given set of
  probability measures.
}
\references{
Christoph Jansen & Georg Schollmeyer & Thomas Augustin 2018\cr
Concepts for Decision Making under Severe
Uncertainty with Partial Ordinal and Partial
Cardinal Preferences.\cr
International Journal of Approximate Reasoning (V. 98)
}
