% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateGIE.R
\name{calculateGIE}
\alias{calculateGIE}
\title{Calculate Generalized Interval Expectation}
\usage{
calculateGIE(ps, delta, p.measures, action)
}
\arguments{
\item{ps}{[\code{PreferenceSystem}]\cr
Preference System. Possibly calculated with
\code{\link{makeDecisionProblem}}, or initialized manually}

\item{delta}{[\code{numeric(1L)}]\cr
granularity parameter. Must be between 0 and 1}

\item{p.measures}{[\code{matrix}]\cr
Matrix of probability vectors. Each row corresponds
to one probability vector.}

\item{action}{[\code{character}]\cr
The act that the interval is calculated for.\cr
Must be one of the levels of the \code{action} variable
in the \code{data.frame} of the object \code{ps$df}.}
}
\value{
[\code{numeric(2L)}]
  Lower and upper bounds of the interval.
}
\description{
Calculates the Generalized Interval Expectation for an act,
  on the basis of a consistent Preference System.
}
\references{
\insertRef{mainref}{dips}
}
