% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatePreferenceSystemConsistency.R
\name{calculatePreferenceSystemConsistency}
\alias{calculatePreferenceSystemConsistency}
\title{Construct A Preference System}
\usage{
calculatePreferenceSystemConsistency(ps)
}
\arguments{
\item{ps}{[\code{dp}]\cr
Decison Problem calculated with \code{\link{makeDecisionProblem}}.}
}
\value{
[\code{list}] With entries:\cr
  R1: Pre order on the acts.\cr
  R2: Preorder on R1.
}
\description{
Constructs a Preference System from a Decision Problem.
}
