% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatePreferenceSystemConsistency.R
\name{calculatePreferenceSystemConsistency}
\alias{calculatePreferenceSystemConsistency}
\title{Calculate Preference System Consistency}
\usage{
calculatePreferenceSystemConsistency(ps)
}
\arguments{
\item{ps}{[\code{dp}]\cr
Preference System calculated with \code{\link{makeDecisionProblem}}.}
}
\value{
[\code{ConsistencyResult}] With entries:\cr
  opt.val: Optimal value of the objective function.\cr
  opt.vec: Optimal found solution vector.
}
\description{
Calculates the granularity up to which the
  given Preference System is consistent.\cr
  Throws an error if computation fails.
}
\references{
\insertRef{mainref}{dips}
}
