% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkPreferenceSystem.R
\name{checkPreferenceSystem}
\alias{checkPreferenceSystem}
\title{Check if an object is a Preference System}
\usage{
checkPreferenceSystem(ps)
}
\arguments{
\item{ps}{[\code{PreferenceSystem}]\cr
Preference System. Possibly calculated with
\code{\link{makeDecisionProblem}}, or initialized manually}
}
\value{
[\code{PreferenceSystem}]
  If checks were successful, the input is returned.
}
\description{
This function is more used internally to check
  if an object has the right structure to qualify
  as a Preference System. However, for users
  that want to set up their own relations R1 and R2
  this function can be useful to check if the
  object was constructed the right way.
  Furthermore, in case the relations \code{R1} and \code{R2}
  are given in form of squared matrices this function will
  convert them to matrices with two and four columns, respectively.
}
