% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatePreferenceSystemConsistency.R
\name{getP}
\alias{getP}
\title{Get Strict Part Of Set}
\usage{
getP(x)
}
\arguments{
\item{x}{[\code{matrix}]\cr
Decison Problem calculated with \code{\link{makeDecisionProblem}}.}
}
\value{
[\code{list}] With entries:\cr
  R1: Pre order on the acts.\cr
  R2: Preorder on R1.
}
\description{
Constructs a Preference System from a Decision Problem.
}
