% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDecisionProblem.R
\name{makeDecisionProblem}
\alias{makeDecisionProblem}
\title{Construct Decision Problem}
\usage{
makeDecisionProblem(df, state, action, exclude = NULL)
}
\arguments{
\item{df}{[\code{data.frame}]\cr
Data from which the Decision Problem shall be derived.}

\item{state}{[\code{character(1L)}]\cr
Name of the variable assigning the states of nature.}

\item{action}{[\code{character(1L)}]\cr
Name of the variable assigning the acts to the data frame rows.}

\item{exclude}{[\code{character}]\cr
Variable names that should be excluded from the decision problem.
Default is \code{NULL}, meaning no variable is excluded.}
}
\value{
[\code{DecisionProblem}] With entries:
  \describe{
    \item{df}{The original data frame. State and action variable names have
    been altered to 'state' and 'action', respectively. Also their
    levels are now given as integer sequence.}
    \item{n.alternatives}{Number of observations in \code{df}}
    \item{ordinal.information}{List of ordinal info for each
      combination of state and action.}
    \item{cardinal.information}{List of ordinal info for each
      combination of state and action.}
    \item{ordinal.vars}{Variable names that provided ordinal information.}
    \item{cardinal.vars}{Variable names that were used to extract the
      cardinal information.}
}
}
\description{
Constructs a Decision Problem from a data frame.
}
