% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGIE.R
\name{plotGIE}
\alias{plotGIE}
\title{Plot Generalized Interval Expectation}
\usage{
plotGIE(ps, deltas, p.measures, actions)
}
\arguments{
\item{ps}{[\code{dp}]\cr
Preference System calculated with \code{\link{makeDecisionProblem}}.}

\item{deltas}{[\code{numeric()}]\cr
Vector of granularity parameters. Each entry must be between 0 and 1.}

\item{p.measures}{[\code{matrix}]\cr
Matrix of probability vectors. Each row corresponds
to one probability vector.}

\item{actions}{[\code{character}]\cr
The acts that the interval is calculated for.\cr
Must be one of the levels of the \code{action} variable
in the \code{data.frame} of the object \code{ps$df}.}
}
\value{
ggplot2 object.
}
\description{
Generates a plot visualizing Generalized Interval Expectation
  for a set of actions from a Preference System
  and a vector of granularities.
}
\references{
Christoph Jansen & Georg Schollmeyer & Thomas Augustin 2018\cr
Concepts for Decision Making under Severe
Uncertainty with Partial Ordinal and Partial
Cardinal Preferences.\cr
International Journal of Approximate Reasoning (V. 98)
}
