% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHasseGraph.R
\name{plotHasseGraph}
\alias{plotHasseGraph}
\title{Plots Hasse-Graph from a Preference System}
\usage{
plotHasseGraph(ps, ...)
}
\arguments{
\item{ps}{[\code{PreferenceSystem}]\cr
Preference System. Possibly calculated with
\code{\link{makeDecisionProblem}}, or initialized manually}

\item{...}{\cr
additional parameters passed down to \code{hasseDiagram::hasse}}
}
\value{
Draws Hasse Graph
}
\description{
Draws a Hasse Graph for the relation \code{R1} stored in
  the Preference System \code{ps}
}
