% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHasseGraph.R
\name{plotHasseGraph}
\alias{plotHasseGraph}
\title{Plot Hasse-Graph from a Preference System}
\usage{
plotHasseGraph(ps, ...)
}
\arguments{
\item{ps}{[\code{dp}]\cr
Preference System calculated with \code{\link{makeDecisionProblem}}.}

\item{...}{\cr
additional parameters passed down to \code{igraph::plot.igraph}}
}
\value{
Draws Hasse Graph
}
\description{
Draws a Hasse Graph for the relation \code{R1} stored in
  the Preference System \code{ps}
}
