% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHasseGraph.R
\name{plotHasseGraph}
\alias{plotHasseGraph}
\title{Plot Hasse-Graph from a Preference System}
\usage{
plotHasseGraph(ps, title = "", ...)
}
\arguments{
\item{ps}{[\code{PreferenceSystem}]\cr
Preference System. Possibly calculated with
\code{\link{makeDecisionProblem}}, or initialized manually}

\item{title}{[\code{character(1L)}]\cr
Title for the plot. Defaults to empty string, meaning no title.}

\item{...}{[any]\cr
Additional parameters passed down to \code{igraph::plot.igraph}.
See \code{\link[igraph]{plot.igraph}} and
\code{\link[igraph]{igraph.plotting}} for details.}
}
\value{
Draws Hasse Graph
}
\description{
Draws a Hasse Graph for the relation \code{R1} stored in
  the Preference System \code{ps}
}
\references{
\insertRef{igraph}{dips}
}
